moved the tree module from the navigator to the application extension
