+++ src/application/browser/index.ts
@@ -11,5 +11,6 @@ export * from './opener-service';
 export * from './application-module';
 export * from './browser';
 export * from './menu';
-export * from './dialogs';
 export * from './widgets';
+export * from './dialogs';
+export * from './tree';
