+++ src/application/browser/style/index.css
@@ -50,4 +50,5 @@ body {
 @import './variables-dark.css';
 @import './materialcolors.css';
 @import './scrollbars.css';
+@import './tree.css';
