+++ src/application/browser/style/tree.css
