+++ src/application/browser/tree/tree-container.ts
