+++ src/application/browser/tree/tree-expansion.ts
