+++ src/application/browser/tree/tree-iterator.ts
