+++ src/application/browser/tree/tree-model.ts
