+++ src/application/browser/tree/tree-widget.ts
