+++ src/navigator/browser/navigator-container.ts
@@ -6,7 +6,7 @@
  */
 
 import { interfaces, Container } from 'inversify';
-import { ITree, ITreeModel, TreeProps, defaultTreeProps } from "./tree";
+import { ITree, ITreeModel, TreeProps, defaultTreeProps } from "../../application/browser";
 import { createFileTreeContainer, FileTree, FileTreeModel, FileTreeWidget, FileTreeServices } from "./file-tree";
 import { FileNavigatorTree } from "./navigator-tree";
 import { FileNavigatorModel, FileNavigatorServices } from "./navigator-model";
